Added the following options:   -q quiet   -l do not recurse   -g globalize %check   -z make %check apply only inside single directories   -u unlink duplications (not by default anymore)
Made value of %check an array of duplicate filepaths, not just the
filepath of the first file with that particular hash; all dups are
added, even if some are unlinked in the process.

Dir's to recurse down into are built up in a temp array and all called
at the end of parsedir() to make -z mangling of %check work properly.